Move peon/peon_defaults out of external vars
Both need to be accessable to all cloud-types, but neither need to
exist in the top-level (highest priority) variables file.  This
file is included by adept as an extra variables file, and will override
all lower priority items.  Instead, define them separetly and just
include them into each cloud-type for the setup context.  This alllows
the defaults to be used but remain easy to override all or in part.

Signed-off-by: Chris Evich <cevich@redhat.com>