fixup! Modify ansible default non-zero exit behavior
When any task fails, ansible skips all other tasks for that host, but
continues with tasks for other hosts.  Because pre-job.xn setup tasks
are mandatory, change the default behavior of ansible to exit
immediatly when any task fails.  The play argument
``any_errors_fatal: true`` accomplishes this.

The exception to the above is during the cleanup transition.  That
context must always complete (successful or not).  In other words,
whether or not kommandir re/discovery is successful, there are still
additional tasks to perform on the exekutir.  Allow those cases
by NOT setting ``any_errors_fatal`` in cleanup_before_job.yml